% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_column_to_fcs.R
\name{add1column}
\alias{add1column}
\title{Add a column to a flowFrame object}
\usage{
add1column(fcs, col, colname = "new data", exprs_in = NULL)
}
\arguments{
\item{fcs}{\code{flowFrame} object.}

\item{col}{numeric vector of values to be appended.}

\item{colname}{name of the appended column.}

\item{exprs_in}{\code{exprs} data matrix to be extended. If NULL, \code{exprs} is extracted from \code{fcs}.}
}
\value{
An extended \code{flowFrame} object.
}
\description{
\code{add1column} appends a column of data to the \code{exprs} data matrix of a \code{flowFrame} object and returns the extended \code{flowFrame} object.
}
