% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08d_Export.R
\name{add_pathways_info_to_FCS}
\alias{add_pathways_info_to_FCS}
\title{Add additional channels to FCS file with info about pathways of interest}
\usage{
add_pathways_info_to_FCS(
  ff,
  pseudotime,
  all_walks,
  trajectories_of_interest,
  fcs_subset_idcs = NULL,
  id
)
}
\arguments{
\item{ff}{\code{flowFrame} object (loaded using package \code{flowCore}) corresponding to FCS with the analysed data}

\item{pseudotime}{list: must contain a numeric vector in the slot \code{res}, corresponding to a pseudotime value for each event included in the analysis}

\item{all_walks}{list of random walks, consisting of node indices in slot \code{v} and start indices referring to entries from \code{v} in \code{starts}}

\item{trajectories_of_interest}{integer vector: walk indices, with entries from \code{1:length(walks$starts)}}

\item{fcs_subset_idcs}{integer vector (optional): indices of selected events
Those are the row numbers of expression matrix from the FCS file which correspond to events included in \code{data}
Default value is \code{1:nrow(data)}, which means all events were selected}

\item{id}{numeric or character: name of the batch of trajectory vertices being appended}
}
\value{
\code{flowFrame} object (corresponds to an enhanced FCS file)
}
\description{
\code{add_pathways_info_to_FCS} requires a \code{flowFrame} object, pseudotime values, all simulated random walks, indices of trajectories of interest,
indices of selected events (in case the analysed data correspond only to a subset of data in the \code{flowFrame}) and a a numeric or character ID for the
specified collection of trajectories.
}
\details{
A valid \code{flowFrame} object with additional channels is returned.
Firstly, the channel \code{'ID_which_event'} contains a binary values tagging which events are included in the trajectories of interest (1000) and which were not (-1).
Secondly, the channel \code{'ID_local_pseudotime'} contains pseudotime values of vertices comprising the trajectories of interest is created.
In both cases \code{'ID'} is replaced with the corresponding parameter value.
You do not need to call this function directly.
}
\references{
\insertRef{Ellis2019}{tviblindi}
}
