% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S3.R
\name{AddLabelsVector.tviblindi}
\alias{AddLabelsVector.tviblindi}
\title{Add a vector of labels per event to a \code{tviblindi} object}
\usage{
\method{AddLabelsVector}{tviblindi}(tv, labels, name = "default")
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{labels}{string or factor vector: label for each event (row in \code{tv$data})}

\item{name}{string: name of the labels vector}
}
\description{
Adds a vector of labels per each row of the \code{data} matrix of a \code{tviblindi} object.
(Multiple alternative label vectors can be supplied, and multiple TI analyses can be conducted, with separate sets of simulated random walks for each labelling of the data.)
}
\details{
The default vector of labels is named '\code{default}'. Use '\code{default}' as a name for your new labels vector to overwrite this.
}
