% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_DimensionReduction.R
\name{AddLayout.tviblindi}
\alias{AddLayout.tviblindi}
\title{Apply dimension reduction to produce a 2-d layout of expression data}
\usage{
\method{AddLayout}{tviblindi}(
  tv,
  method = "vaevictis",
  name = "default",
  layout = NULL,
  n_epochs = if (method == "vaevictis") 100L else if (method == "umap") 200 else NULL,
  n_neighbours = if (is.null(tv$kNN)) 50L else ncol(tv$kNN$idcs) - 1,
  load_model = NULL
)
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{method}{string: dimension-reduction method. One of \code{vaevictis} (encoder with t-SNE regularisation), \code{umap} or \code{tsne}. Default value is \code{vaevictis}}

\item{name}{string: name of layout to be generated. Default value is \code{default}}

\item{layout}{optional numeric matrix with number of rows equal to \code{nrow(tv$data)}: pre-computed layout matrix, to be added instead of creating a new layout. Default value is \code{NULL}}

\item{n_epochs}{integer: maximum number of epochs for training a \code{vaevictis} or \code{umap} model. Default is \code{100} for \code{vaevictis} and 200 for \code{umap}}

\item{n_neighbours}{integer: neighbour count for \code{umap}. Default value is \code{50}}

\item{load_model}{character vector of length 2: paths to files created by \code{tv$vae$save(file1, file2)} for loading a \code{vaevictis} model instead of training a new one.
Default value is \code{NULL}, which causes a new model to be trained}

\item{vsplit}{number: portion of data used as validation set in \code{vaevictis} method. Default value is \code{0.1}}

\item{encoder_shape}{integer vector of length 2: shape (depth and width) of encoder in \code{vaevictis}. Default value is \code{c(128,128,128)}}

\item{decoder_shape}{integer vector of length 2: shape (depth and width) of decoder in \code{vaevictis}. Default value is \code{c(128,128,128)}}

\item{perplexity}{number: perplexity parameter for t-SNE regularisation in \code{vaevictis} or in t-SNE. Default value is \code{10}}

\item{batch_size}{integer: batch size for \code{vaevictis} training. Default value is \code{512}}

\item{patience}{integer: maximum patience parameter for \code{vaevictis} training (early stopping). Default value is \code{0}}

\item{alpha}{number: penalty parameter for t-SNE regularisation in \code{vaevictis}. Default value is \code{10}}
}
\description{
Apply dimension reduction to produce a 2-d layout of expression data
}
\references{
\insertRef{Nl2008}{tviblindi}

\insertRef{McInnes2018}{tviblindi}

\insertRef{Ushey2020}{tviblindi}
}
