% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Filtration.R
\name{alpha_complex_filtration_values}
\alias{alpha_complex_filtration_values}
\title{Compute alpha-complex filtration values}
\usage{
alpha_complex_filtration_values(filtration, coordinates)
}
\arguments{
\item{filtration}{list: filtration of a simplicial complex. Must contain definition (vertex indices) of each simplex by indices in \code{cmplx}. \code{values} slot (with filtration values) is disregarded, as filtration values must are re-computed from scratch}

\item{coordinates}{numeric matrix of coordinates: rows of matrix correspond to vertices, number of columns is number of dimensions}
}
\value{
list: new filtration, with \code{cmplx} (simplex indices) and \code{values} (filtration values for each simplex)
}
\description{
Provided a simplicial complex filtration and corresponding point coordinates, this function calculates the corresponding filtration values for each simplex.
\code{filtration} must be a list containing the slot \code{cmplx}, which is a list of integer vectors: 1-based vertex indices.
These define simplices in the alpha-complex filtration.
}
