% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Pseudotime.R
\name{AssignDistances.tviblindi}
\alias{AssignDistances.tviblindi}
\title{Assign distance-from-origin to each vertex for expression data in \code{tviblindi}}
\usage{
\method{AssignDistances}{tviblindi}(
  tv,
  name,
  origin_name,
  max_analytical = 1000,
  n_iter = 1500,
  tolerated_error = 1e-06
)
}
\arguments{
\item{tv}{\code{tviblindi}-class object with \code{kNN}, \code{dist} and \code{trans}}

\item{name}{string: name of transition model to use}

\item{origin_name}{string or integer: name of cell-of-origin to use}

\item{max_analytical}{integer: maximum number of vertices in data (without origin) for which we compute an exact solution. Default value is \code{1000}}

\item{tolerated_error}{number: error tolerance for conjugate gradient method. Default value is \code{1e-6}}

\item{nb_it}{integer: maximum number of iterations for conjugate gradient method. Default value is \code{1500}}
}
\description{
This method adds a new slot \code{pseudotime} to an object of class \code{tviblindi}.
}
\details{
Distance from cell-of-origin to each vertex is computed using an existing transition matrix.
Solving a system of linear equations, the average number of transitions needed to reach each vertex in \code{data} is computed.
If the number of vertices in \code{data} is less than a pre-set threshold, an analytical (exact) solution is computed; otherwise we apply conjugate gradient to solve numerically.
}
\references{
\insertRef{Bates2019}{tviblindi}

\insertRef{Bates2014}{tviblindi}

\insertRef{Eddelbuettel2013}{tviblindi}
}
