% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Filtration.R
\name{boundary_matrix}
\alias{boundary_matrix}
\title{Compute boundaries of a filtration}
\usage{
boundary_matrix(filtration)
}
\arguments{
\item{filtration}{list: filtration of a simplicial complex. Must contain description of each simplex by indices in \code{cmplx} and associated filtration values in \code{values}}
}
\value{
list with slots \code{cmplx} (boundary simplex indices) and \code{values} (filtration values for each simplex)
}
\description{
Provided a simplicial complex filtration, \code{boundary_matrix} returns a list of boundary simplices (\code{cmplx}) and unmodified filtration values (\code{values}).
\code{filtration} must be a list containing definition (vertex indices) of each simplex by indices in \code{cmplx} and filtration values of each simplex in \code{values}.
}
