% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_boundary_C.R
\name{build_boundaryC}
\alias{build_boundaryC}
\title{Generate boundary matrix object}
\usage{
build_boundaryC(dFilt)
}
\arguments{
\item{dFilt}{list; filtration of a simplicial complex. Contains description of each simplex by indices in \code{cmplx} and filtration values in \code{values}.}
}
\value{
A list of \code{cmplx} (boundaries) and \code{values} (same as in \code{dFilt}).
}
\description{
Provided a list of simplices, \code{build_boundaryC} makes list of preceding simplices which are boundaries. This is, effectively, a compact representation of a boundary matrix.
}
