% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Cluster.tviblindi}
\alias{Cluster.tviblindi}
\title{Computes k-means clusters for triangulation, modifies x}
\usage{
\method{Cluster}{tviblindi}(
  x,
  K = 625,
  method = "kmeans",
  kmeans_algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  clustering = NULL
)
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default 625); centers for kmeans clustering}

\item{method}{"som" or "kmeans" (default)}

\item{clustering}{integer vector of the same length as \code{nrow(x$data)}. User provided landmark clusters}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Cluster}
}
