% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Connectome.tviblindi}
\alias{Connectome.tviblindi}
\title{Create connectome from tvilblindi class object}
\usage{
\method{Connectome}{tviblindi}(x, ...)
}
\arguments{
\item{x}{tviblindi class object.}

\item{png}{chracter; path to the png file to plot connectome. If NULL, connectome is plotted directly. Optimised for png}

\item{K}{integer (default K=30); number of nearest neighbors for louvain clustering}

\item{origin_name}{integer/character (default 1); path model to use}

\item{labels}{integer/character (default 1); labels to use}

\item{layout}{integer/character (default 1); layout to use}

\item{clusters}{integer vector (default NULL); provide clustering externally}

\item{arrow.sizefactor}{numeric vector (default 1); adjust size of an arrow}

\item{legend.position}{(default "topright")}

\item{lcex}{numeric( default 1); legend cex}

\item{notplot}{character (character vector) (default "ungated"); which populations should not be shown in piecharts}

\item{qq}{numeric (default 0); only arrows of edges width above qq percentile will be plotted}

\item{directed}{boolean (default TRUE); plot only edges of major direction}
}
\value{
\code{tviblindi} returns an invisible tviblindi class object.
}
\description{
\code{Connectome}
}
\details{
Computes louvain clusters and estimates the flow (and the direction) between them from simulated walks
(parameter \code{equinumerous} in \code{Walks}) would bias the result!
}
