% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_tools.R
\name{consensus_path}
\alias{consensus_path}
\title{Generate consensus path}
\usage{
consensus_path(pathways, class, coordinates)
}
\arguments{
\item{pathways}{a pathways list object, as generated by \code{create_pathways_object}.}

\item{class}{a single pathway class number by which to filter pathways.}

\item{coordinates}{coordinate matrix, preferrably of non-reduced dimensionality, to use for averaging pathways.}
}
\value{
\code{consensus_path} returns a coordinate matrix containing vertices of consensus path.
}
\description{
\code{consensus_path} takes a set of pathways (generated from simulations of random walks through a graph) and creates a consensus path which represents
 their topology. Length of the consensus path is 3 times the max walk length. Missing points are interpolated.
}
