% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Graphs.R
\name{ConstructkNNG.tviblindi}
\alias{ConstructkNNG.tviblindi}
\title{Compute a k-nearest-neighbours matrix}
\usage{
\method{ConstructkNNG}{tviblindi}(tv, k = 100, approximate = TRUE, n_trees = 150)
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{k}{integer: nearest neighbour count}

\item{approximate}{logical: whether to apply an approximate algorithm (faster). Default value is \code{TRUE}}

\item{n_trees}{integer: number of look-up trees to be constructed by \code{annoy}, if applicable. Default value is \code{150}}
}
\description{
This method adds a new slot \code{kNN} to an object of class \code{tviblindi}.
For exact k-NN, a multi-threaded version of ball-tree nearest neighbour search is used.
For the approximate version, the \code{annoy} algorithm is used.
}
\details{
A list \code{kNN} with slots \code{idcs} and \code{dist}, containing matrices with nearest neighbour indices and distances (respectively), with each row corresponding to a row of \code{data}, is produced.
}
\references{
\insertRef{Nl2008}{tviblindi}

\insertRef{Ulyanov2016}{tviblindi}

\insertRef{Bernhardsson2020}{tviblindi}
}
