% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08a_Triangulation.R
\name{contract_walks}
\alias{contract_walks}
\title{Contract random walks using result of clustering}
\usage{
contract_walks(walks, clusters)
}
\arguments{
\item{walks}{list of random walks, consisting of node indices in slot \code{v} and start indices referring to entries from \code{v} in \code{starts}}

\item{clusters}{integer vector: assignments of each node along the walk to a cluster}
}
\value{
modified \code{walks} list
}
\description{
\code{contract_walks} re-creates random walks through a graph using assignments of nodes to clusters.
}
