% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Copy.tviblindi}
\alias{Copy.tviblindi}
\title{Deep copy of x}
\usage{
\method{Copy}{tviblindi}(x)
}
\arguments{
\item{x}{tviblindi class object.}
}
\value{
returns a tviblindi class object.
}
\description{
\code{Copy}
}
