% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Filtration.R
\name{create_k_skeleton}
\alias{create_k_skeleton}
\title{Create k-skeleton of a simplicial complex}
\usage{
create_k_skeleton(filtration, coordinates, k, type = "alpha")
}
\arguments{
\item{filtration}{list: filtration of a simplicial complex. Must contain definition (vertex indices) of each simplex by indices in \code{cmplx}. \code{values} slot (with filtration values) is disregarded, as filtration values must are re-computed from scratch}

\item{coordinates}{numeric matrix of coordinates: rows of matrix correspond to vertices, number of columns is number of dimensions}

\item{k}{integer: skeleton \code{k} (dimensionality)}

\item{type}{character: type of filtration. Only allowed value now is '\code{alpha}'}
}
\value{
list; new filtration, with \code{cmplx} (simplex indices) and \code{values} (filtration values for each simplex)
}
\description{
Provided a simplicial complex filtration and corresponding point coordinates, this function  generates a \code{k}-skeleton of the complex.
}
