% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_tools.R
\name{create_pathways_object}
\alias{create_pathways_object}
\title{Create a pathways list object}
\usage{
create_pathways_object(walks, classes)
}
\arguments{
\item{walks}{a \code{walks} list object, as generated by \code{random_walk_N} or related function.}

\item{classes}{a vector of class numbers corresponding to walks.}
}
\value{
\code{create_pathways_object} returns a list with slots \code{walks} and \code{classes}.
}
\description{
\code{create_pathways_object} takes a \code{walks} list object, as generated by \code{random_walk_N} or related function, and
 generates a pathways list object, containing a \code{walks} slot with distinct walks (index vectors) and a \code{classes} slot
 with class numbers per walk.
}
