% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Denoise.tviblindi}
\alias{Denoise.tviblindi}
\title{Reduce noise in data for the purpose of triangulation, modifies x}
\usage{
\method{Denoise}{tviblindi}(x, K = 30, iter = 1, method = "KNN", kernel = "SEMer", sym = "max")
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default K=30); number of neigbors to average (see details).}

\item{iter}{integer (default 1); number of iterations (see details).}

\item{method}{character; either "KNN" or "MAGIC"}

\item{kernel}{character; for method=="MAGIC", see \code{knn.adj2spadjsim}.}

\item{sym}{character (default "max"); for method=="MAGIC", how to symmetrize the transition matrix (options "none", "max", "min", "mean", "prob").}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Denoise}
}
\details{
A simple noise reducution algorithm is applied - every cell coordinates are replaced by the average of \code{K} nearest neigbors.
This process is repeated \code{iter}-times.
}
