% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Downsampling.R
\name{Downsample}
\alias{Downsample}
\title{Create downsampled version of tviblindi object}
\usage{
Downsample(
  tv,
  N = 10000,
  k = 10,
  method = "density",
  intrinsic_dim = 2,
  labels_name = "default",
  threshold = floor(N/100),
  ignore_labels = NULL
)
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{N}{integer: size of the domwnsampled data. Default value is \code{10000}}

\item{k}{integer: number of nearest neighbors to estimate density if density-based downsampling is used. Default value is \code{10}}

\item{method}{string: name of downsampling method to apply. Either \code{density} (for density-based downsampling) or \code{labels} (for downsampling biased against the removal of small labelled populations).
Default value is \code{density}}

\item{intrinsic_dim}{integer: expected intrinsic dimension of data. (Higher dimension under-estimates the density of sparse regions.) Default value is \code{2}}

\item{labels_name}{string: name of labels vector to use by method \code{labels}. Default value is '\code{default}'}

\item{threshold}{integer: threshold value for minimum population size in method \code{labels}. Default value is \code{floor(N / 100)}}

\item{ignore_labels}{string or string vector: names of labelled populations that should be left out entirely from sampling, for method \code{labels}. Default value is \code{NULL}}
}
\description{
This method takes a \code{tviblindi} object and creates a modified version with downsampled expression data.
}
\details{
However, working with the full data (without downsampling), if possible, is encouraged.

Density-based downsampling (for \code{method == 'density'}) should reduce over-abundant dense populations.
Label-based downsampling (for \code{method == 'labels'}) prevents the vanishing of smaller populations while downsampling, and works with a previously specified vector of population labels.
}
