% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Filtration.tviblindi}
\alias{Filtration.tviblindi}
\title{Computes triangulation, boundary matrix and reduced boundary matrix, modifies x}
\usage{
\method{Filtration}{tviblindi}(x, method = "witness", K = 30, alpha2 = NULL)
}
\arguments{
\item{x}{tviblindi class object.}

\item{method}{character (only "witness" complex is implemeted); Uses Gudhi and CGAL libraries to compute witness complex}

\item{K}{integer (default K=30); number of nearest neighbors.}

\item{alpha}{double (default \code{NULL}); relaxation parameter for witness complex. If \code{NULL} mean distance to
K-th nearest witness is used.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Filtration}
}
