% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{get_rep}
\alias{get_rep}
\title{Get representation cycle representation in terms of simplicial complexes}
\usage{
get_rep(cycle, boundary)
}
\arguments{
\item{cycle}{n-simplex indices forming a cyclical n-chain.}

\item{boundary}{reduced boundary matrix object, as generated by \code{reduce_boundary}.}
}
\value{
\code{get_rep} returns a vector of column indices of \code{boundary}.
}
\description{
\code{get_rep} generates representation of a cycle in terms of complexes created during filtration, using a reduced boundary matrix object, as generated
 by \code{reduce_boundary}. The sum of these complexes (effectively a symmetric difference) is the cycle.
}
