% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09a_ShinyLauncher.R
\name{Interactive}
\alias{Interactive}
\title{Launch interactive Shiny interface for tviblindi}
\usage{
Interactive(tviblindi_s3, show_all_fates = TRUE)
}
\arguments{
\item{tviblindi_s3}{\code{tviblindi} class object}

\item{show_all_fates}{logical: whether all potential terminal nodes (according to the transition model) should be shown, regardless of whether any simulated walks actually terminate in them}
}
\description{
WARNING: running this function creates a folder names \code{tviblindi_tmp} in the working directory.
}
\references{
\insertRef{Chang2019}{tviblindi}

\insertRef{Pederson2019}{tviblindi}

\insertRef{Perrier2020}{tviblindi}

\insertRef{Bailey2015}{tviblindi}

\insertRef{DeVries2016}{tviblindi}

\insertRef{Galili2015}{tviblindi}

\insertRef{Wickham2019}{tviblindi}

\insertRef{Ellis2019}{tviblindi}

\insertRef{Wickham2019a}{tviblindi}

\insertRef{Warnes2020}{tviblindi}

\insertRef{Kratochvil2020}{tviblindi}
}
