% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{knn.adj.construct}
\alias{knn.adj.construct}
\title{Create a sparse adjacency matrix object}
\usage{
knn.adj.construct(XX, K, metric = "euclid")
}
\arguments{
\item{XX}{matrix of coordinates. Rows correspond to vertices.}

\item{K}{number of nearest neighbours.}

\item{metric}{distance metric. \code{euclid} or \code{cosine}.}
}
\value{
\code{knn.adj.construct} returns a matrix consisting of 3 rows. 1st row contains (1-based) vertex indices. 2nd row contains indices of nearest neighbours. 3rd row contains distances between data points in rows 1 and 2.
}
\description{
\code{knn.adj.construct} generates an adjacency matrix object for a K-nearest-neighbours graph. Nearest neighbour indices and neighbour distances are found. Nearest neighbour search is implemented via a vantage-point tree algorithm (used in Van der Maaten, 2014).
}
\references{
Van der Maaten, L (2014). Accelerating t-SNE using Tree-Based Algorithms. Journal of Machine Learning Research 15 (2014) 1-21
}
