% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{knn.adj2spadjsim}
\alias{knn.adj2spadjsim}
\title{Convert sparse adjacency matrix to similarity matrix}
\usage{
knn.adj2spadjsim(adj, kernel = c("Exp"), epsilon = NULL)
}
\arguments{
\item{adj}{sparse adjacency matrix, as generated by \code{knn.raw2adj}.}

\item{kernel}{kernel function to use for computing similarity.}
}
\value{
A ``sparseMatrix`` object containing similarities of each pair of points from the original adjacency matrix.
}
\description{
\code{knn.adj2spadjsim} converts a sparse adjacency matrix, as generated by \code{knn.raw2adj}, to a similarity matrix using some kernel function.
 Distinctly from an adjacency matrix, entries in a similarity matrix are dependent upon, but not equivalent to, distances.
 They are calculated using a probability density function (PDF). Standard deviation of distribution is taken as median distance or some function thereof.
 The particular distance value is supplied as the quantile parameter for the PDF.
}
\examples{
# create an artificial data set of two clusters in 2-dimensional space
coord <- cbind(c(rnorm(1000,100,20), rnorm(500, 20, 5)), c(rnorm(1000,100,20), rnorm(500, 120, 5)))

sim <- knn.spadj2sym(  #             to symmetric
  knn.adj2spadjsim(    #          to similarity matrix
   knn.raw2adj(        #       to sparse
     knn.adj.raw(      #    raw adjacency matrix
       coords, K=5))))

}
