% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{knn.diffuseX}
\alias{knn.diffuseX}
\title{Generate diffusion matrix from coordinates}
\usage{
knn.diffuseX(X, K = 30, neigen = dim(X)[2], metric = "euclid")
}
\arguments{
\item{X}{matrix of coordinates. Rows correspond to vertices.}

\item{K}{number of neighbours to use in construction a kNN graph.}

\item{neigen}{number of eigenvectors and eigenvalues to be computed. Defaults to number of dimensions.}

\item{metric}{distance metric. \code{euclid} or \code{cosine}.}
}
\value{
\code{sparse.diffuse} returns a list containing the following elements.

\code{X} is a matrix of diffusion coordinates.

\code{phi0} is the left eigenvector representing
stationary distribution of the diffusion process.

\code{eigenvals} is a vector containing eigenvalues of \code{Asp}, except for the first (largest) value, corresponding to \code{phi0}.

\code{psi} is a matrix of right eigenvectors as rows.

\code{phi} is a matrix of left eigenvectors as columns.

\code{neigen} is the input parameter \code{neigen}.

\code{eigenmult} are eigen-multipliers of the diffusion map.
}
\description{
\code{knn.diffuseX} generates a diffusion matrix. Diffusion coordinates are extracted via eigendecomposition of a generalised Laplacian of a K-nearest-neighbours graph.
 Adapted from package \code{diffusionMap}.
}
