% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{knn.raw2adj}
\alias{knn.raw2adj}
\title{Convert adjacency matrix to a sparseMatrix object}
\usage{
knn.raw2adj(adj)
}
\arguments{
\item{adjraw}{adjacency matrix.}
}
\value{
\code{knn.raw2adj} returns a matrix consisting of 3 rows. 1st row contains (1-based) vertex indices. 2nd row contains indices of nearest neighbours. 3rd row contains distances between data points in rows 1 and 2.
}
\description{
\code{knn.raw2adj} converts an adjacency matrix, as generated by \code{knn.adj.raw} to a \code{sparseMatrix} object, as generated by \code{knn.adj.construct}.
}
\examples{
# create an artificial data set of two clusters in 2-dimensional space
coord <- cbind(c(rnorm(1000,100,20), rnorm(500, 20, 5)), c(rnorm(1000,100,20), rnorm(500, 120, 5)))

coord.adj <- knn.adj.raw(coord, K=5)
coord.adj.sparse <- knn.raw2adj(coord.adj)

}
