% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{knn.spadj2sym}
\alias{knn.spadj2sym}
\title{Symmetrise similarity (or adjacency) matrix}
\usage{
knn.spadj2sym(adj)
}
\arguments{
\item{adj}{similarity or adjacency matrix (\code{sparseMatrix} object).}
}
\value{
Symmetrised matrix (\code{sparseMatrix} object).
}
\description{
\code{knn.spadj2sym} converts a similarity matrix, as generated by, for instance, \code{knn.adj2spadjsim}, (or, optionally, an adjacency matrix) to a symmetric version thereof.
 For each corresponding pair of entries (row \code{i}, column \code{j} and row \code{j}, column \code{i}), we take the maximum of either entry.
}
\examples{
# create an artificial data set of two clusters in 2-dimensional space
coord <- cbind(c(rnorm(1000,100,20), rnorm(500, 20, 5)), c(rnorm(1000,100,20), rnorm(500, 120, 5)))

sim <- knn.adj2spadjsim(     #         to similarity matrix
         knn.raw2adj(        #      to sparse
           knn.adj.raw(      #   raw adjacency matrix
             coord, K=5)))   # 5 nearest neighbours

isSymmetric(sim) # FALSE

sim <- knn.spadj2sym(sim)

isSymmetric(sim) # TRUE

}
