% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{KNN.tviblindi}
\alias{KNN.tviblindi}
\title{Computes KNN matrix, modifies x}
\usage{
\method{KNN}{tviblindi}(x, K = 100, method = "annoy", trees = 150, denoised = FALSE)
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default K=100); number of nearest neighbors.}

\item{method}{character ("BT" or "annoy";default "annoy"); implements either ball tree nearest neigbor search (https://github.com/lvdmaaten/bhtsne)
or Approximate Nearest Neighbors Oh Yeah (https://github.com/spotify/annoy).}

\item{trees}{integer (default 150); number of trees for annoy - more trees more precision and more time of computation.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{KNN}
}
