% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Graphs.R
\name{kNN_annoy}
\alias{kNN_annoy}
\title{Construct a k-NN matrix using the 'annoy' algorithm}
\usage{
kNN_annoy(data, k, trees = 150)
}
\arguments{
\item{data}{numeric matrix: coordinates (one data point per each row)}

\item{k}{integer: nearest neighbour count}

\item{trees}{integer: number of look-up trees to generate by \code{annoy}. The higher the number of trees, the higher the accuracy. Default value is \code{150}}
}
\value{
a list \code{kNN} with slots \code{idcs} and \code{dist} containing matrices with nearest neighbour indices and distances, respectively, with each row corresponding to a row in \code{data}
}
\description{
\code{kNN_annoy} takes a coordinate matrix \code{data}, a nearest neighbour count \code{k} and a look-up tree count \code{trees}.
It finds the \code{k} nearest neighbours to each vertex defined in \code{data} using an approximate algorithm.
}
\references{
\insertRef{Nl2008}{tviblindi}

\insertRef{Ulyanov2016}{tviblindi}

\insertRef{Bernhardsson2020}{tviblindi}
}
