% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_launcher.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launches shiny interface}
\usage{
launch_shiny(tviblindi_s3, ShowAllFates = TRUE)
}
\arguments{
\item{tviblindi_s3}{tviblindi class object.}
}
\description{
\code{Filtration}
}
\details{
See \code{ToggleShowFates}.
}
