% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_registry.R
\name{list_dimred_methods}
\alias{list_dimred_methods}
\title{List registered dimensionality-reduction methods}
\usage{
list_dimred_methods()
}
\value{
character vector of method names usable as \code{DimRed(x, method=...)}.
}
\description{
List registered dimensionality-reduction methods
}
