% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_Pseudotime.R
\name{ModelTransitions.tviblindi}
\alias{ModelTransitions.tviblindi}
\title{Compute transition model for expression data in a \code{tviblindi} object}
\usage{
\method{ModelTransitions}{tviblindi}(tv, name, origin_name, k, kernel = "Exp", epsilon = NULL)
}
\arguments{
\item{tv}{\code{tviblindi}-class object with \code{kNN}}

\item{name}{string: name of transition model to generate}

\item{origin_name}{string: name of cell-of-origin to use}

\item{k}{integer: number of nearest neighbours to use for calculating transition probabilities (upper bound is \code{k} of \code{tv$kNN})}

\item{kernel}{string: name of kernel function to use.
One of \code{Exp} (exponential), \code{SE} (standard error), \code{Lap} (Laplacian). Default value is \code{Exp}}

\item{epsilon}{numeric: kernel function parameter (epsilon). Default value is \code{NULL}, whereby a good epsilon is estimated automatically}
}
\description{
This method adds a new slot \code{trans} to an object of class \code{tviblindi}, containing transition probabilities associated with graph edges.
A chosen kernel is applied to compute transition probabilities from distances in \code{kNN}, with parameter \code{epsilon}.
}
\references{
\insertRef{Bates2019}{tviblindi}
}
