% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{path_triangulation}
\alias{path_triangulation}
\title{Generate triangulated paths}
\usage{
path_triangulation(walks, X, cmplx, sel)
}
\arguments{
\item{walks}{object specifying walks through the graph containing vertices, as specified by \code{X}. As generated by \code{random_walk} or of the same format.}

\item{X}{matrix of vertex coordinates together with indices mapping), with rows corresponding to vertices. Result of \code{select_unique_points}}

\item{cmplx}{simplicial complex, as generated by a filtration function, such as \code{alphaComplexFiltration}.}

\item{sel}{vector containing indices of selected walks.}
}
\value{
\code{path_triangulation} returns a list of point (vertex) indices.
}
\description{
\code{path_triangulation} re-creates random walks, as generated by \code{random_walk}, using 1-simplices generated by a filtration function.
 Thus, the original paths are converted into 1-chains
 which approximate them and represent their topology. Since all vertices from the original path must be included, the triangulated path must be of same or greater length.
 Connections between each pair of points are found in terms of 1-simplices via uniform-cost search.
}
