% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{pathways.visual.labels.segments}
\alias{pathways.visual.labels.segments}
\title{Visualise pathways progression: cell population labels by segment}
\usage{
pathways.visual.labels.segments(paths, classes, which.class, labels,
  pseudotime, breaks = NULL, n.part = 20, exp.part = 1,
  proportional = TRUE)
}
\arguments{
\item{paths}{a list of pathways. This corresponds to the \code{walks} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{classes}{a list of classes corresponding to pathways. This corresponds to the \code{classes} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{which.class}{number of the pathway class used for generating the plot.}

\item{labels}{cell population labels (eg. gates) corresponding to single events (to which \code{paths} refer).}

\item{pseudotime}{a pseudotime value vector, corresponding to vertices of \code{coordinates}.}

\item{breaks}{if not null, this overrides \code{n.part} and \code{exp.part}: create pseudotime partitions by dividing the
\code{pseudotime} vector by its values (instead of including a fixed number of cells in each segment). \code{breaks} can
consist either of a single values (number of uniform partitions) or of a vector of breaks (cut-off values) between 0 and 1.}

\item{n.part}{number of partitions per pseudotime. Used which \code{breaks} is null. A fixed number of cells, picked from a
series sorted by pseudotime values, is included in each partition.}

\item{exp.part}{used with \code{n.part} if \code{breaks} is null. Non-negative exponent used in computing the series of cut-off values for
pseudotime vector partitioning. If equal to 1, partitioning is uniform (each parition contains an equal number of cells). If
greater than one, partition size shrinks with increasing pseudotime (recommended, useful if many mature cells/terminal events are present in sample).
If between 0 and 1, partition size expands with increasing pseudotime.}

\item{proportional}{Boolean. If \code{TRUE}, show relative representations of cell populations. If \code{FALSE}, show absolute cell counts.
Default to \code{TRUE}.}
}
\value{
\code{pathways.visual.labels.segments} returns a \code{ggplot} object.
}
\description{
\code{pathways.visual.parameters.segments} generates a \code{ggplot2} visualisation of differential representation
 of labelled cell populations with pseudotime progression, summing out pathways from a single class over pseudotime segments.
}
