% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{pathways.visual.parameters.all_data}
\alias{pathways.visual.parameters.all_data}
\title{Visualise pathways progression: single parameter using all pathway points}
\usage{
pathways.visual.parameters.all_data(paths, classes, which.class,
  coordinates, pseudotime, which.param, expression.transform = NULL,
  asinh.denominator = 5)
}
\arguments{
\item{paths}{a list of pathways. This corresponds to the \code{walks} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{classes}{a list of classes corresponding to pathways. This corresponds to the \code{classes} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{which.class}{number of the pathway class used for generating the plot.}

\item{coordinates}{a coordinate matrix to which the input pathways refer. This is a matrix of expression values, column correspond to
parameters (markers) and rows correspond to single events.}

\item{pseudotime}{a pseudotime value vector, corresponding to vertices of \code{coordinates}.}

\item{which.param}{name of the parameter (from \code{coordinates} column names) used for generating the plot.}

\item{expression.transform}{transformation to apply on expression values (entries of \code{coordinates}). Accepts \code{"asinh"}.
Defaults to null (no transformation).}

\item{asinh.denominator}{if \code{expression.transform} is \code{"asinh"}, which factor \code{F} to use in \code{coordinates <- asinh(coordinates/F)}.
\code{F = 5} recommended for CyTOF data, approximately \code{F = 120} for flow data.}
}
\value{
\code{pathways.visual.parameters.segments} returns a \code{ggplot} object.
}
\description{
\code{pathways.visual.parameters.all_data} generates a \code{ggplot2} visualisation of differential expression
 for a single parameter with pseudotime progression, showing all pathways coming from a single class.
}
