% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{pathways.visual.parameters.segments}
\alias{pathways.visual.parameters.segments}
\title{Visualise pathways progression: single parameter by segments}
\usage{
pathways.visual.parameters.segments(paths, classes, which.class,
  coordinates, pseudotime, which.param, breaks = NULL, n.part = 10,
  exp.part = 1, expression.transform = NULL, asinh.denominator = 5)
}
\arguments{
\item{paths}{a list of pathways. This corresponds to the \code{walks} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{classes}{a list of classes corresponding to pathways. This corresponds to the \code{classes} slot of a pathways list object, as generated
by \code{create_pathways_object}.}

\item{which.class}{number of the pathway class used for generating the plot.}

\item{coordinates}{a coordinate matrix to which the input pathways refer. This is a matrix of expression values, column correspond to
parameters (markers) and rows correspond to single events.}

\item{pseudotime}{a pseudotime value vector, corresponding to vertices of \code{coordinates}.}

\item{which.param}{name of the parameter (from \code{coordinates} column names) used for generating the plot.}

\item{breaks}{if not null, this overrides \code{n.part} and \code{exp.part}: create pseudotime partitions by dividing the
\code{pseudotime} vector by its values (instead of including a fixed number of cells in each segment). \code{breaks} can
consist either of a single values (number of uniform partitions) or of a vector of breaks (cut-off values) between 0 and 1.}

\item{n.part}{number of partitions per pseudotime. Used which \code{breaks} is null. A fixed number of cells, picked from a
series sorted by pseudotime values, is included in each partition.}

\item{exp.part}{used with \code{n.part} if \code{breaks} is null. Non-negative exponent used in computing the series of cut-off values for
pseudotime vector partitioning. If equal to 1, partitioning is uniform (each parition contains an equal number of cells). If
greater than one, partition size shrinks with increasing pseudotime (recommended, useful if many mature cells/terminal events are present in sample).
If between 0 and 1, partition size expands with increasing pseudotime.}

\item{expression.transform}{transformation to apply on expression values (entries of \code{coordinates}). Accepts \code{"asinh"}.
Defaults to null (no transformation).}

\item{asinh.denominator}{if \code{expression.transform} is \code{"asinh"}, which factor \code{F} to use in \code{coordinates <- asinh(coordinates/F)}.
\code{F = 5} recommended for CyTOF data, approximately \code{F = 120} for flow data.}
}
\value{
\code{pathways.visual.parameters.segments} returns a \code{ggplot} object.
}
\description{
\code{pathways.visual.parameters.segments} generates a \code{ggplot2} visualisation of differential expression
 for a single parameter with pseudotime progression, grouped by pathways coming from a single class. Pseudotime
 progression is divided into segments for which mean expression per pathway is computed and plotted.
}
