% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{pers_diagram}
\alias{pers_diagram}
\title{Create persistence diagram}
\usage{
pers_diagram(dBr, plot = TRUE, repre = NULL)
}
\arguments{
\item{dBr}{reduced boundary matrix object, as generated by \code{reduce_boundary}.}

\item{plot}{Boolean, plot persistence diagram? Defaults to \code{TRUE}.}
}
\value{
\code{pers_diagram} returns a list containing the following elements.

\code{inds} is a matrix describing homology classes by dimension, index of simplex which caused birth of the class and index of simplex which caused death of the class.

Analogically, \code{vals} describes homology classes by dimension, filtration value corresponding to birth and filtration value corresponding to death.
}
\description{
\code{pers_diagram} computes persistence and creates a persistence diagram given a reduced boundary matrix object, as generated by
 \code{reduce_boundary}.
}
