% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S3.R
\name{Plot.tviblindi}
\alias{Plot.tviblindi}
\title{Plot 2-dimensional layout of data}
\usage{
\method{Plot}{tviblindi}(
  tv,
  layout_idx = 1,
  labels_idx = 1,
  transition_model_idx = 1,
  markers = NULL,
  labels = NULL,
  pch = ".",
  cex = 3,
  col = "labels",
  plots_per_row = 4,
  main = NULL,
  cex.main = NULL,
  hide_legend = FALSE,
  ...
)
}
\arguments{
\item{tv}{\code{tviblindi}-class object with \code{layout}}

\item{layout_idx}{integer or string: index of layout to use. Default value is \code{1}}

\item{labels_idx}{integer or string: index of per-event label vector to use. Default value is \code{1}}

\item{transition_model_idx}{integer or string; index of transition model to use. Default value is \code{1}}

\item{markers}{string or string vector: name of marker of interest for constructing a heatmap. If multiple marker names are given (maximum 32), a grid of plots is produced.  If a substring of a marker name is given and it can be matched, the most probable match will be used. Default value is \code{NULL}}

\item{labels}{string or string vector: name of population of interest for constructing a 2-d map of that population. If multiple population names are given (maximum 32), a grid of plots is produced. If parameter \code{markers} is not \code{NULL}, it overrides this parameter. Default value is \code{NULL}}

\item{pch}{integer or character: \code{pch} parameter of the \code{plot} method (plotting character or symbol). Default value is '\code{.}' (a dot)}

\item{cex}{numeric: \code{cex} parameter of the \code{plot} method: plotting character or symbol size. Default value is \code{3}}

\item{col}{string: can be '\code{labels}' or '\code{pseudotime}' to specify type of plot. '\code{labels}' is for plotting a map of all labelled cell populations; '\code{markers}' is for showing a grid of heatmaps for the expression of each marker. If either parameter \code{markers} or \code{labels} is not \code{NULL}, it overrides this parameter. Default value is '\code{labels}'.}

\item{plots_per_row}{integer: if a grid of multiple plots is to be generated, how many plots should fit onto a single row? Default value is \code{4}}

\item{main}{string: if specified, this overrides the default \code{main} parameter of \code{plot} (title of the plot, or absence thereof). Default value is \code{NULL}}

\item{cex.main}{number: if specified, this overrides the default \code{cex.main} parameter of \code{plot} (font size of the plot title). Default value is \code{NULL}}
}
\description{
Plots previously generated 2-dimensional layout of expression data in \code{tviblindi} object, with custom colour-coding of data points.
}
\details{
To colour events by pseudotime, make sure the \code{pseudotime} slot is filled first. A yellow-to-brown-to-red (increasing) colour range is used to denote pseudotime values.

Colouring events by labels gives a plot with a colour-coded legend.
Alternatively (if the previous approach results in an overly cluttered plot), he localisation of point belonging to a single label of interest can be displayed.

Finally, yellow-to-brown-to-red heat map of expression of a single marker of interest can be displayed
}
