% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{plot_walks}
\alias{plot_walks}
\title{Plot simulated random walks}
\usage{
plot_walks(
  XX,
  walks,
  sel = 1:length(walks$starts),
  col = 1:length(walks$starts),
  Xmap = NULL,
  ...
)
}
\arguments{
\item{XX}{matrix of vertex coordinates, with rows corresponding to vertices.}

\item{walks}{object specifying walks through the graph containing vertices, as specified by \code{XX}. As generated by \code{random_walk}.}

\item{sel}{vector containing indices of selected walks to be plotted. Defaults to all walks.}

\item{col}{vector of numeric colour codes for plotting. Defaults to numbers of each walk in the \code{walks} object.}

\item{Xmap}{optional vector which maps numbers in \code{walks$v} to row indices of \code{XX}.}

\item{...}{additional parameters passed to \code{lines} function used for plotting walks.}
}
\description{
\code{plot_walks} plots simulated random walks, provided coordinates of vertices and walks through a graph containing the vertices. These walks are typically generated by the function \code{random_walk}.
}
