% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Pseudotime.tviblindi}
\alias{Pseudotime.tviblindi}
\title{Computes pseudotime, modifies x}
\usage{
\method{Pseudotime}{tviblindi}(x, K = 30)
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default K=30); number of nearest neighbors to compute transition matrix.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Pseudotime}
}
\details{
Computes average distance of each cell from the cell-of-origin of all random walks in undirected graph of nearest neigbors.
}
