% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk_N}
\alias{random_walk_N}
\title{Generate random walks through graph}
\usage{
random_walk_N(A, origin, N)
}
\arguments{
\item{A}{transition matrix.}

\item{origin}{index of the origin vertex.}

\item{N}{number of walks to simulate.}
}
\value{
\code{random_walk_N} returns a list containing the following elements. \code{v} is an integer vector of all simulated walks in terms of vertex
 indices. \code{starts} is an integer vector containing indices of \code{v} which correspond to start-points of each simulated walk.
}
\description{
\code{random_walk_N} generates a set of random walks through a graph, provided an oriented transition matrix, index of origin vertex (starting point
 for all paths) and number of walks to simulate. Since we are working with a connected oriented graph, there exist end-vertices where a walk must
 terminate. Therefore, we simply use entries in the transition matrix to calculate probability of transitioning between vertices and simulate a pre-set
 number of walks iteratively.
}
