% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RandomWalks.R
\name{random_walks}
\alias{random_walks}
\title{Simulate random walks}
\usage{
random_walks(transition_matrix, origin, N)
}
\arguments{
\item{transition_matrix}{a SparseMatrix object; matrix of transition probabilities between indexed nodes of a graph.}

\item{origin}{index of the node which is taken as origin of each walk.}

\item{N}{number of walks to simulate.}
}
\value{
a list with the following slots.
\code{v} is an integer vector with the sequences of node indices comprising each walk, concatenated back-to-back.
\code{starts} is an integer vector of indices in \code{v} that denotes which entries in \code{v} correspond to the first nodes in walks.
}
\description{
Given a set of oriented edges associated with transition probabilities, \code{random_walks} simulates random walks through a graph as 1st-order Markov processes.
You do not need to call this method directly.
}
