% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{reduce_boundary}
\alias{reduce_boundary}
\title{Generate a reduced boundary matrix object}
\usage{
reduce_boundary(boundary)
}
\arguments{
\item{boundary}{boundary matrix object, as generated by \code{build_boundaryC}.}
}
\value{
\code{reduce_boundary} returns a list containing the following elements.

\code{boundary} is a list indexed by simplex numbers, containing vectors of respective boundary simplices.

\code{nonzero_col} is a vector containing indices of non-zero columns of a reduced boundary matrix.

\code{low} is a vector containing indices of the lowest non-zero entry row for all non-zero columns of the reduced boundary matrix.

\code{dim} is a vector containing the dimension of each complex represented by a non-zero column.

\code{values} is a vector of filtration values.
}
\description{
\code{reduce_boundary} generates a compact representation of a reduced boundary matrix, produced from a boundary matrix object, as generated by \code{build_boundaryC}.
}
