% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_Filtration.R
\name{reduced_boundary_matrix}
\alias{reduced_boundary_matrix}
\title{Reduce boundary matrix of a filtration}
\usage{
reduced_boundary_matrix(boundary)
}
\arguments{
\item{boundary}{list containing the slot \code{cmplx}, which is a list of vectors (1-based simplex indices), and the slot \code{values} containing filtration values of each simplex}
}
\value{
list with slots \code{boundary}, \code{nonzero_col}, \code{low} and \code{dim}.
}
\description{
Provided a boundary matrix corresponding to a simplicial complex filtration, \code{reduced_boundary_matrix} returns a list with slots \code{boundary} (mapping of simplices to their boundaries), \code{nonzero_col}, \code{low} and \code{dim} (see reference below for description of the algorithm).
}
\references{
\insertRef{Edelsbrunner2008}{tviblindi}
}
