% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{remove_back_R}
\alias{remove_back_R}
\title{Remove backward edges from "triples" matrix}
\usage{
remove_back_R(triplets, time)
}
\arguments{
\item{triplets}{edge matrix, as generated by \code{sparse2triples}.}

\item{time}{distance assignment, as generated by \code{assign_distance}.}
}
\value{
\code{remove_back_R} outputs the \code{triplets} matrix, possibly with some rows removed.
}
\description{
\code{remove_back_R} takes a "triples" matrix of edges, as generated by \code{sparse2triples} and \code{time}, as generated by \code{assign_distance}.
 Edges such that the \code{from} vertex is assigned a distance value by \code{time} lower or equal to value assigned to the \code{to} vertex are removed.
}
\examples{
# create an artificial data set of two clusters in 2-dimensional space
coord <- make_snowman(1000)
origin <- 1

# create a similarity matrix
sim <- knn.spadj2sym(  #             to symmetric
  knn.adj2spadjsim(    #          to similarity matrix
   knn.raw2adj(        #       to sparse
     knn.adj.raw(      #    raw adjacency matrix
       coords, K=5))))

pseudotime <- assign_distance(sim, origin) # assign pseudotime
oriented <- remove_back_R(sparse2triples(sim), # change matrix format
                            pseudotime$res)      # orient the graph according to pseudotime values

}
