% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_RandomWalks.R
\name{remove_cycles}
\alias{remove_cycles}
\title{Remove cycles from simulated random walks}
\usage{
remove_cycles(walks)
}
\arguments{
\item{walks}{list of random walks, consisting of node indices in slot \code{v} and start indices referring to entries from \code{v} in \code{starts}}
}
\value{
modified \code{walks} without cycles
}
\description{
\code{remove_cycles} requires a list with the folowing slots: \code{v} is an integer vector with the sequences of node indices comprising each walk,
concatenated back-to-back; \code{starts} is an integer vector of indices in \code{v} that denotes which entries in \code{v} correspond to the first nodes in walks.
The function removes any cycles within the walks.
You do not need to call this function directly.
}
\references{
\insertRef{Eddelbuettel2014}{tviblindi}

\insertRef{Bates2014}{tviblindi}

\insertRef{Eddelbuettel2013}{tviblindi}
}
