% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_tools.R
\name{select_paths_points}
\alias{select_paths_points}
\title{Select a subset of path points}
\usage{
select_paths_points(paths, sel)
}
\arguments{
\item{paths}{output of \code{random_walk}.}

\item{sel}{vector of numbers of walks to be selected.}
}
\value{
\code{select_paths_points} returns a vector of vertex indices in selected walks.
}
\description{
\code{select_paths_points} returns a subset of points on walks from an object generated by \code{random_walk}.
}
