% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hompath.R
\name{select_unique_points}
\alias{select_unique_points}
\title{Select unique points}
\usage{
select_unique_points(XX, sel)
}
\arguments{
\item{XX}{matrix of vertex coordinates, with rows corresponding to vertices.}

\item{sel}{vector of points (i.e. row indices of \code{XX}).}
}
\value{
\code{select_unqiue_points} returns a list containing the following elements. \code{X} is the original matrix of coordinates (\code{XX}). \code{map}
 is a numeric vector containing indices of unique entries in \code{sel}.
}
\description{
\code{select_unique_points} requires a vector of points (i.e. vertex indices) and a coordinate matrix. The function returns a list including
 indices of unique points and original coordinates.
}
