% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08a_Triangulation.R
\name{select_walks_nodes}
\alias{select_walks_nodes}
\title{Select nodes of random walks by walk indices}
\usage{
select_walks_nodes(walks, selection)
}
\arguments{
\item{walks}{list of random walks, consisting of node indices in slot \code{v} and start indices referring to entries from \code{v} in \code{starts}}

\item{selection}{integer vector: walk indices, with entries from \code{1:length(walks$starts)}.}
}
\value{
modified \code{walks} list
}
\description{
\code{select_walks_nodes} returns a modified \code{walks} list, containing only those walks which are selected by their indices in the original \code{walks} object.
}
