% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Graphs.R
\name{SetOrigin.tviblindi}
\alias{SetOrigin.tviblindi}
\title{Set cell-of-origin}
\usage{
\method{SetOrigin}{tviblindi}(tv, label, name = "default")
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{label}{integer or string: either label of population of origin or index of the cell-of-origin}

\item{name}{string: name for the particular cell-of-origin. Default value is '\code{default}'}
}
\description{
This method adds a new cell-of-origin to a \code{tviblindi} object.
A point with the chosen index is set as cell-of-origin, in which simulated random walks start.
Alternatively, a point in the data is found such that it is closest to the centroid of a chosen annotated population and picked as a cell-of-origin.
}
