% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Som.tviblindi}
\alias{Som.tviblindi}
\title{Computes SOM clusters for triangulation, modifies x}
\usage{
\method{Som}{tviblindi}(x, xdim = 25, ydim = 25)
}
\arguments{
\item{x}{tviblindi class object.}

\item{xdim, }{ydim integer (default 25); SOM mesh size}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Som}
}
