% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{sparse.Laplacian.construct}
\alias{sparse.Laplacian.construct}
\title{Generate a sparse Laplacian matrix from an adjacency matrix}
\usage{
sparse.Laplacian.construct(adj, kernel = c("Exp"), epsilon = NULL)
}
\arguments{
\item{adj}{adjacency matrix, as generated by \code{knn.raw2adj}.}

\item{kernel}{kernel function for calculating transition probabilities: \code{Exp} for exponential kernel, \code{SE} for SE kernel or \code{Lap} for Laplace kernel. Defaults to \code{Exp}.}
}
\value{
A \code{sparseMatrix} object representing a symmetric graph Laplacian.
}
\description{
\code{sparse.Laplacian.construct} produces a generalised graph Laplacian, which can be seen as a symmetrised transition matrix. Transition probabilities are computed using a specified kernel function.
}
