% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvilblindi_core.R
\name{sparse2triples}
\alias{sparse2triples}
\title{Convert sparse similarity matrix to "triples" matrix of edges}
\usage{
sparse2triples(m)
}
\arguments{
\item{m}{similarity (or analogous) matrix (\code{sparseMatrix} object).}
}
\value{
\code{sparse2triples} generates a 3-column matrix. Columns correspond to \code{from} vertex, \code{to} vertex and edge \code{weight}, respectively.
}
\description{
\code{sparse2triples} converts a \code{sparseMatrix} object (typically, a similarity matrix) to a "triples" format: a matrix describing oriented graph edges.
 Columns correspond to vertices and edge weight.
}
\examples{
# create an artificial data set of two clusters in 2-dimensional space
coord <- make_snowman(1000)
origin <- 1

# create a similarity matrix
sim <- knn.spadj2sym(  #             to symmetric
  knn.adj2spadjsim(    #          to similarity matrix
   knn.raw2adj(        #       to sparse
     knn.adj.raw(      #    raw adjacency matrix
       coords, K=5))))

pseudotime <- assign_distance(sim, origin) # assign pseudotime
oriented <- remove_back_R(sparse2triples(sim), # change matrix format
                            pseudotime$res)      # orient the graph according to pseudotime values

}
