% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{ToggleShowFates.tviblindi}
\alias{ToggleShowFates.tviblindi}
\title{Helper method to set ShowAllFates, modifies x}
\usage{
\method{ToggleShowFates}{tviblindi}(x, ShowAllFates = NULL)
}
\arguments{
\item{x}{tviblindi class object.}

\item{ShowAllFates;}{if NULL x$ShowAllFates<-!x$ShowAllFates}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{ToggleShowFates}
}
\details{
If \code{x$ShowAllFates==TRUE} all potential fates (see \code{Walks}) are displayed in shiny app, otherwise only ends
of actual simulations are displayed. \code{x$ShowAllFates==FALSE} by default.
}
