% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08a_Triangulation.R
\name{triangulate_trajectory}
\alias{triangulate_trajectory}
\title{Generate triangulated trajectory}
\usage{
triangulate_trajectory(points, cluster_codes, cmplx_hashed, graph)
}
\arguments{
\item{points}{integer vector: indices of the nodes which comprise a random walk (vertices of the original point cloud)}

\item{cluster_codes}{numeric matrix: cluster-center coordinates, with each row corresponding to a single cluster center}

\item{cmplx_hashed}{a hashed environment with the simplices in a filtration, as generated by \code{tviblindi:::hash_cmplx}.}

\item{graph}{an \code{igraph} object representing the graph in which to run the shortest paths search (implemented in \code{tviblindi:::.update_walks_by_termini})}
}
\value{
integer vector of graph node indices
}
\description{
\code{triangulate_trajectory} re-creates a random walk usign 1-simplices generated by a filtration function, Dijkstra's shortest paths search.
The function outputs a 1-chain which approximates and represent their topology.
Since all vertices from the original walk must be included, the triangulated trajectory must be of the same or of greater length.
}
\references{
\insertRef{Csardi2006}{tviblindi}

\insertRef{Eddelbuettel2013}{tviblindi}
}
