% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_dyntoydata.R
\docType{data}
\name{tviblindi_dyntoydata}
\alias{tviblindi_dyntoydata}
\title{dyntoy datatset for tviblindi}
\format{
An object of class \code{"data.frame"}; First column group_id 
character vector; columns 2:26 first 25 PCs
}
\usage{
data(tviblindi_dyntoydata)
}
\description{
dyntoy dataset is created as follows, principle component analisis is 
performed and first 25 PCs is used.
}
\details{
devtools::install_github('dynverse/dyntoy')

n_events <- 10000 \cr
n_features <- 3896 \cr
set.seed(12345) \cr
d <- dyntoy::generate_dataset( \cr
id           = 'tviblindi_dyntoy_test', \cr
  model        = 'connected', \cr
  num_features = n_features, \cr
  num_cells    = n_events \cr
)
}
\examples{
data(tviblindi_dyntoydata)
group_id<-tviblindi_dyntoydata[,1]
datainput<-as.matrix(tviblindi_dyntoydata[,-1])
tv1<-tviblindi(data=datainput,labels=group_id)
DimRed(tv1,epochs=5) # few epochs to keep the example quick; default is 100
DimRed(tv1,method="umap")

Set_origin(tv1,label = "M4",origin_name = "M4_hitting_time")
Set_origin(tv1,label = "M4",origin_name = "M4_hitting_distance")
KNN(tv1)
Cluster(tv1,K=225) #kmeans clustering
Filtration(tv1) #default setting is very conservative, less simplices could be created with same resolution (e.g. Filtration(tv1,alpha2=1))

Pseudotime(tv1,weighted = FALSE,origin_name = "M4_hitting_time")
Walks(tv1,N=1000,origin_name = "M4_hitting_time", K=15)

Pseudotime(tv1,weighted = TRUE,origin_name = "M4_hitting_distance")
Walks(tv1,N=1000,origin_name = "M4_hitting_distance", K=15)
launch_shiny(tv1)
}
\references{
Cannoodt, R., and W. Saelens, 2022: dyntoy: Generating simple toy
data of cellular differentiation. R package version 0.9.9.
(\href{github.com/dynverse/dyntoy}{dyntoy})
}
\keyword{datasets}
