% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Upsampling.R
\name{Upsample}
\alias{Upsample}
\title{Create a new \code{tviblindi} object with upsampled data}
\usage{
Upsample(tv, n = 5, k = 30, idx_labels = 1, idcs = NULL, concentration = 0.1)
}
\arguments{
\item{tv}{\code{tviblindi}-class object with a k-nearest-neighbour graph computed}

\item{n}{integer: how many points shall be sampled for each upsampled point in original data. Default value is 5}

\item{k}{integer: number of nearest neighbours of each original point to use for fitting the Dirichlet distribution to draw new point from. Default value is 30}

\item{idx_labels}{integer or string: index of the vector of labels to use for re-labelling the data. Default value is 1}

\item{idcs}{integer or string vector: row indices of points in \code{tv$data} to which upsampling shall be applied, or a vector of labels of these points. Default value is \code{NULL} (this reverts to all points in \code{tv$data})}

\item{concentration}{float: alpha parameter of the Dirichlet distribution. Default value is 0.1}
}
\description{
\code{Upsample} creates a new \code{tviblindi} object, taking the associated input expression matrix with its labels and sampling new additional points.
Dirichlet distribution over neighbours of each point is used to sample the new points.
Any slots besides \code{data} and \code{labels} will be cleared by this.
}
