% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Clustering.R
\name{Cluster.tviblindi}
\alias{Cluster.tviblindi}
\title{Compute k-means or SOM clusters for triangulation of data}
\usage{
\method{Cluster}{tviblindi}(
  tv,
  method = "kmeans",
  kmeans_algorithm = "Hartigan-Wong",
  k = 625,
  xdim = 25,
  ydim = 25
)
}
\arguments{
\item{tv}{\code{tviblindi}-class object with \code{kNN}}

\item{method}{string: clustering method specification, either '\code{som}' or '\code{kmeans}'. Default value is '\code{k-means}'}

\item{kmeans_algorithm}{string: algorithm for k-means clustering. One of '\code{Hartigan-Wong}', '\code{Lloyd}', '\code{Forgy}', '\code{MacQueen}'. Default value is '\code{Hartigan-Wong}'}

\item{k}{integer: cluster count for k-means clustering. Default value is 625}

\item{xdim}{integer: width of Kohonen map computed by \code{FlowSOM} (if \code{method == 'som'}). Default value is 25}

\item{ydim}{integer: height of Kohonen map computed by \code{FlowSOM} (if \code{method == 'som'}). Default value is 25}
}
\description{
This method adds new slots \code{clusters} and \code{codes} to an object of class \code{tviblindi}.
These contain (1) the mapping of vertices in \code{data} onto nodes in a newly trained k-means model or self-organising map (SOM) and (2) cluster centers, respectively.
}
\references{
\insertRef{VanGassen2015}{tviblindi}
}
