% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectome.R
\name{connectome}
\alias{connectome}
\title{Create connectome from tvilblindi class object}
\usage{
connectome(x, png = "connectome.png", K = 70)
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default K=30); number of nearest neighbors for louvain clustering}
}
\value{
\code{tviblindi} returns an invisible tviblindi class object.
}
\description{
\code{Connectome}
}
\details{
Computes louvain clusters and estimates the flow (and the direction) between them from simulated walks 
(parameter \code{equinumerous} in \code{Walks}) would bias the result!
}
