% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Connectome}
\alias{Connectome}
\title{Computes connectome, modifies x}
\usage{
Connectome(x, ...)
}
\arguments{
\item{x}{tviblindi class object.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Pseudotime}
}
\details{
See \code{connectome}
}
