% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_Connectome.R
\name{Connectome}
\alias{Connectome}
\title{Create connectome-analysis object from a tviblindi object}
\usage{
Connectome(
  tv,
  k = 30,
  layout_name = "default",
  labels_name = "default",
  transition_model_name = "default",
  png = NULL
)
}
\arguments{
\item{tv}{trajectory-inference object of class \code{tviblindi}}

\item{k}{integer: number of nearest neighbors for Louvain clustering}

\item{layout_name}{string: name of 2-dimensional layout to use for connectome analysis}

\item{labels_name}{string: name of labels vector used to display pie charts for each connectome graph node}

\item{transition_model_name}{string: name of transition model to use for connectome analysis}

\item{png}{optional string: file path to a PNG file of the diagram to be generated. Default value is \code{NULL}, whereby the currently active plotting device is used instead}
}
\description{
Generates a diagram of connections between Louvain clusters of high-dimensional expression data, indicating flow (and direction) between tentative developmental stages.
This is based on previously simulated random walks (stochastic traversals of a k-NN graph governed by properties of a previously generated transition model).
If \code{equinumerous} was set to true in random walks (equal number of walks per terminus enforced), this biases the results (undesirable).
}
