% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract_walks.R, R/path_tools.R
\name{contract_walks}
\alias{contract_walks}
\title{Contract walks to cluster nodes}
\usage{
contract_walks(walks, clusters)

contract_walks(walks, clusters)
}
\arguments{
\item{walks}{output of \code{random_walk}.}

\item{cluster}{vector of cluster assignement of the original pointcloud.}

\item{walks}{output of \code{random_walk}.}

\item{cluster}{vector of cluster assignement of the original pointcloud.}
}
\value{
\code{contract_wlaks} returns a list of vertex indices and indices of walks starts.

\code{contract_wlaks} returns a list of vertex indices and indices of walks starts.
}
\description{
\code{contract_walks} returns a walks with vericies contracted to cluster nodes from an object generated by \code{random_walk}.

\code{contract_walks} returns a walks with vericies contracted to cluster nodes from an object generated by \code{random_walk}.
}
