% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_vals.R
\name{create_k_skeleton}
\alias{create_k_skeleton}
\title{Create k-skeleton of a simplicial complex}
\usage{
create_k_skeleton(filtration, coordinates, k, type = "alpha",
  threshold = 0)
}
\arguments{
\item{filtration}{list; filtration of a simplicial complex. Must contain description of each simplex by indices in \code{cmplx}. \code{values} slot is disregarded, as filtration values must be computed from scratch.}

\item{coordinates}{matrix of coordinates. Rows of matrix correspond to vertices, columns correspond to dimensions.}

\item{k}{skeleton dimensionality.}

\item{type}{type of filtration: 'alpha' or 'rips'. Default is 'alpha'.}

\item{threshold}{Rips threshold, used it type is 'rips'.}
}
\value{
list; new filtration, with \code{cmplx} (simplex indices) and \code{values} (filtration values for each simplex).
}
\description{
Provided a simplicial complex filtration and corresponding point coordinates, \code{create_k_skeleton} generates a skeleton of the given complex in \code{k} dimensions.
}
