% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Graphs.R
\name{Denoise.tviblindi}
\alias{Denoise.tviblindi}
\title{Reduce noise in high-dimensional biological data in a \code{tviblindi} object}
\usage{
\method{Denoise}{tviblindi}(tv, k = 30, n_iter = 1)
}
\arguments{
\item{tv}{\code{tviblindi}-class object}

\item{k}{integer: number of neighbours to use for averaging the coordinates. Default value is \code{30}}

\item{n_iter}{integer: number of iterations. Default value is \code{1}}
}
\description{
This method adds a denoised expression matrix to a \code{tviblindi} object.
Each vertex' coordinates are replaced by the average of its nearest neighbours' coordinates.
This is repeated \code{n_iter} times.
This modification hopefully brings the data points closer to the original manifold and removes technical noise.
}
