% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Denoise.tviblindi}
\alias{Denoise.tviblindi}
\title{Reduce noise in data for the purpose of triangulation, modifies x}
\usage{
\method{Denoise}{tviblindi}(x, K = 30, iter = 1)
}
\arguments{
\item{x}{tviblindi class object.}

\item{K}{integer (default K=30); number of neigbors to average (see details).}

\item{iter}{integer (default 1); number of iterations (see details).}
}
\value{
\code{tviblindi} returns an invisible tviblindi class object.
}
\description{
\code{Denoise}
}
\details{
A simple noise reducution algorithm is applied - every cell coordinates are replaced by the average of \code{K} nearest neigbors. 
This process is repeated \code{iter}-times.
}
