% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{DimRed.tviblindi}
\alias{DimRed.tviblindi}
\title{Dimensional reduction, modifies x}
\usage{
\method{DimRed}{tviblindi}(
  x,
  method = c("vaevictis", "diffuse"),
  layout = NULL,
  dim = 2,
  vsplit = 0.1,
  enc_shape = c(128, 128, 128),
  dec_shape = c(128, 128, 128),
  perplexity = 10,
  batch_size = 512L,
  epochs = 100L,
  patience = 0L,
  ivis_pretrain = 0,
  ww = c(10, 10, 1, 1),
  margin = 1,
  shuffle = FALSE,
  neigen = 2,
  t = 0,
  K = 30,
  load_model = NULL,
  upsample = NULL,
  labels_name = names(x$labels)[1]
)
}
\arguments{
\item{x}{tviblindi class object.}

\item{method}{character; "vaevictis" - deep (auto)encoder (combination of ideas from different papers - to be eleborated)
or "diffuse" - saparse diffusion maps}

\item{layout}{numeric matrix \code{[nrow(data),2]} (optional); if \code{!is.null(layout)} dimensional reduction is plugged-in.}

\item{dim}{integer (default 2); dimension of reduced data (for the moment only resonable value is 2 but the reduction should work for
any integer value (for "diffuse" \code{< nrow(data)})).}

\item{vsplit}{double (default 0.1); percentage of data used as validation step in "vaevictis".}

\item{enc_shape}{integer vector (default \code{c(128,128,128)}); shape (depth and wisth) of the encoder.}

\item{dec_shape}{integer vector (default \code{c(128,128,128)}); shape (depth and wisth) of decoder.}

\item{perplexity}{double (default 10.); perplexity for tsne regularisation see https://www.nature.com/articles/s41467-018-04368-5.}

\item{batch_size}{integer (default 512); batch size for "vaevictis" training.}

\item{epochs}{integer; maximum number of epochs for "vaevictis" training.}

\item{patience}{integer; maxim patience for for "vaevictis" training (early stopping).}

\item{ww}{vector double; weights for vaevictis in this order - tsne_regularisations, ivis pn loss, reconstruction error, KL divergence}

\item{margin}{double; ivis pn loss margin}

\item{shuffle}{logical; shuffle data before validation split; involves recomputation of KNN matrix}

\item{neigen}{integer; for "diffuse" number of eigen vectors to compute.}

\item{t}{double; time parameter for "diffuse", if \code{t==0} multi-time scale is used (geometric sum).}

\item{load_model}{character vector of 2 components; paths to files created by by x$vae$save(file1,file2) - model is loaded and applied}

\item{upsample}{named list \code{list(N=,takeall=)} or \code{NULL};  sample events by labels (involves recomputation of KNN matrix); affects "vaevictis" only; if NULL nothing happens, \code{N} events per label, takes all events from labels in character vector \code{takeall}}

\item{labels_name}{name of label vector if one is used for upsampling and \code{x} has mutliple label vectors.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{DimRed}
}
\details{
The pathway analysis visualisation benefits from dimensional reductions which are by definition continuous... to be elaborated
}
