% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{DownSample.tviblindi}
\alias{DownSample.tviblindi}
\title{Downsamples data, modifies x}
\usage{
\method{DownSample}{tviblindi}(x, N = 10000, K = 10, method = "default", e = 1, D = 2)
}
\arguments{
\item{x}{tviblindi class object.}

\item{N}{integer; size of the domwnsampled data.}

\item{K}{integer; number of nearest neighbors to estimate density.}

\item{method}{character; deprecated.}

\item{e}{double; deprecated.}

\item{D}{integer; expected intrinsic dimension of the data - higher dimension underestimates the density of sparse regions.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{DownSample}
}
\details{
Density-based downsampling of the data, should reduce over-abundant dense populations. The labels, events_sel, data and layout are downsampled, 
the rest is set to \code{NULL}
}
