% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_launcher.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch graphical user interface for tviblindi trajectories analysis}
\usage{
launch_shiny(input_fcs_path, origin, pseudotime, coords, coords_clusters,
  clusters, filtration, layout, walks_raw, term, b, rb)
}
\arguments{
\item{input_fcs_path}{Path to analysed FCS file.}

\item{origin}{Index in *coords* of first point in all trajectories.}

\item{pseudotime}{Pseudotime object, as generated by *assign_distance*.}

\item{coords}{Full expression matrix.}

\item{coords_clusters}{Clustered expression matrix.}

\item{clusters}{Cluster assignments (per event).}

\item{filtration}{Filtration on all points in *coords*.}

\item{layout}{2D projection matrix.}

\item{walks_raw}{Walks object as output of function *randow_walk_...*.}

\item{term}{Terminal node indices per walk.}

\item{b}{Boundary matrix of *filtration*.}

\item{rb}{Reduced boundary matrix of *filtration*.}
}
\description{
Launch graphical user interface for tviblindi trajectories analysis
}
