% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cycles.R
\name{remove_cycles}
\alias{remove_cycles}
\title{Remove cycles from paths}
\usage{
remove_cycles(walks, verbose = TRUE, detailed = FALSE)
}
\arguments{
\item{walks}{a \code{walks} object, as generated by \code{random_walk_N} or related function.}

\item{verbose}{Boolean; if true, report how many trajectories contained cycles.}

\item{detailed}{Boolean; if true and if verbose is true, report which trajectories contained cycles.}
}
\value{
\code{remove_cycles} returns a modified \code{walks} object.
}
\description{
\code{remove_cycles} takes a \code{walks} object (simulations of random walks through a graph) and removes any
 cycles occuring within the walks.
}
