% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{Set_origin.tviblindi}
\alias{Set_origin.tviblindi}
\title{Sets cell-of-origin, modifies x}
\usage{
\method{Set_origin}{tviblindi}(x, label)
}
\arguments{
\item{x}{tviblindi class object}

\item{label}{character or integer; either label of population of origin (cell nearest to the mean of the population will be consider as origin)
or index of cell-of-origin.}
}
\value{
returns an invisible tviblindi class object.
}
\description{
\code{Set_origin}
}
