% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_S3.R
\name{tviblindi}
\alias{tviblindi}
\title{Constructor of a tviblindi-analysis object}
\usage{
tviblindi(
  data,
  labels,
  fcs_path = NULL,
  fcs_subset_idcs = NULL,
  analysis_name = NULL
)
}
\arguments{
\item{data}{numeric matrix: pre-processed protein expression or read count data matrix. Column names shall correspond to name of individual markers (detected on a transctipt or protein level)}

\item{labels}{string or factor vector of the same length as row count of \code{data}. Assigns a label to each measured event, mapping the event to a known cell population. Unassigned cells should be labelled as '\code{ungated}' or '\code{\*ungated\*}'
If \code{labels} is left empty, all events are assigned the '\code{ungated}' label}

\item{fcs_path}{optional string: path to FCS file from which cytometry expression \code{data} was extracted. The use of \code{fcs_path} is encouraged if a subset of expression data from a known FCS file is used. If \code{fcs_path} is available, an enhanced version of the FCS file (with information about detected developmental pathways) can be generated for downstream analysis in FlowJo. Default value is \code{NULL}}

\item{fcs_subset_idcs}{integer vector: row numbers in the expression matrix of an associated FCS file that correspond to rows of \code{data}.
(Default value is \code{NULL}, denoting that all events from the associated FCS file, if any such file exists, were taken along for the trajectory-inference analysis)}

\item{analysis_name}{optional string: name of the trajectory-inference analysis. If not specified, a name is created by concatenation of user-name, date, time and \code{tviblindi} version}
}
\value{
\code{tviblindi} returns an object of class \code{tviblindi}
}
\description{
Constructs an object of class \code{tviblindi} for trajectory-inference analysis.
This object allows for semi-automated detection of developmental trajectories in expression data acquired via high-dimensional cytometry, scRNA-seq or CITE-seq ADT data.
}
\details{
After calling this constructor, use the methods \code{SetOrigin}, \code{ConstructkNNG}, \code{Denoise}, \code{AddLayout}, \code{Cluster}, \code{Filter}, \code{ComputePseudotime}, \code{SimulateRandomWalks} and \code{Interactive} to conduct you trajetory-inference analysis.

Alternatively, use the function \code{InferTrajectories} instead of using this constructor and these methods to run the whole \code{tviblindi} pipeline.
}
