% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tviblindi_S3.R
\name{tviblindi}
\alias{tviblindi}
\title{Constructor of tviblindi class}
\usage{
tviblindi(
  data,
  labels,
  fcs_path = NULL,
  events_sel = NULL,
  keep_intermediate = FALSE
)
}
\arguments{
\item{data}{a numeric matrix; the (transformed and compensated) expression data.}

\item{labels}{character or factor vector of the same length as \code{nrow(data)};
assignment of cells to population, unassigned cells should be labelled as "ungated" for better plotting.}

\item{fcs_path}{character (optional); path to fcs file to store the results - typically the fcs file with analysed data.}

\item{events_sel}{integer vector (optional); the indices of data in fcs file, defaults to 'code{1:nor(data)}.}

\item{keep_intermediate}{bool (default FALSE); if intermediate matrices (transition probability and spar distance matrix)}

\item{ShowAllFate}{bool; if all teoretical fates (see \code{Walks}) should be stored and plotted
should be kept during computation}
}
\value{
\code{tviblindi} returns a tviblindi class object.
}
\description{
\code{tviblindi}
}
